requirements.txt
django
django-allauth
djangorestframework
jwt
openai
python-docx
requests